feat: transition to RV64I architecture
- Changed all registers and addresses from u32 to u64
- Added read64/write64 to memory and bus
- Updated LW to read32 with sign extension (RV64 behavior)
- Changed shift operations shamt from 5-bit to 6-bit
- Updated all tests for 64-bit expected values

Steps 1-3 and 6 of RV64 transition complete

+++ src/devices/uart.rs
@@ -3,8 +3,8 @@ use std::{
     io::{Write, stdout},
 };
 
-pub const UART_BASE: u32 = 0x10000000;
-pub const UART_SIZE: u32 = 8;
+pub const UART_BASE: u64 = 0x10000000;
+pub const UART_SIZE: u64 = 8;
 
 const UART_RBR: u32 = 0; // Receive Buffer Register
 const UART_THR: u32 = 0; // Transmit Holding Register
